+++ mulog-core/dev/user.clj
@@ -139,6 +139,83 @@
 
 
 
+
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+;;                                                                            ;;
+;;                ----==| O P E N   T E L E M E T R Y |==----                 ;;
+;;                                                                            ;;
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+
+
+(comment
+
+  (def st
+    (u/start-publisher!
+      {:type :console :pretty? true}))
+
+  (def st2
+    (u/start-publisher!
+      {:type :zipkin :url "http://localhost:9411/"}))
+
+  (comment
+    (require '[com.brunobonacci.mulog.publishers.open-telemetry :as ot])
+
+    (def otp (ot/open-telemetry-publisher
+               {:type :open-telemetry
+                :url  "http://localhost:4318/"}))
+
+    (def st2
+      (u/start-publisher! {:type :inline :publisher otp}))
+    )
+
+  (u/set-global-context! {:app "demo" :version 1 :env "local"})
+
+  (u/log :test :t (rand))
+
+  (u/trace :test-trace
+    [:foo 1, :t (rand)]
+    (Thread/sleep (rand-int 50)))
+
+  (u/trace :test-trace-wth-result
+    {:pairs [:foo 1, :t (rand)] :capture #(select-keys % [:hello])}
+    {:hello "world" :capture "test"})
+
+  (u/trace :test-trace-capture-error
+    {:pairs [:foo 1, :t (rand)] :capture #(select-keys % [:hello])}
+    (rand-int 100))
+
+  (u/trace :test-trace-wth-result
+    {:pairs [:foo 1, :t (rand)] :capture (fn [x] {:return x})}
+    (rand-int 100))
+
+  (u/trace :test-syntax-error
+    (identity {:pairs [:foo 1, :t (rand)] :capture-result :hello})
+    {:hello "world"})
+
+  (u/trace :big-operation
+    [:v 1 :level 0]
+    (Thread/sleep (rand-int 2000))
+
+    (u/trace :small-operation
+      [:level 1 :seq 1]
+      (Thread/sleep (rand-int 2000)))
+
+    (u/trace :small-operation
+      [:level 1 :seq 2]
+      (Thread/sleep (rand-int 2000))
+
+      (u/trace :operation
+        [:level 3]
+        (Thread/sleep (rand-int 1000)))
+
+      (Thread/sleep (rand-int 10)))
+    (Thread/sleep (rand-int 200)))
+
+  (st2)
+  (st)
+  )
+
+
 ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
 ;;                                                                            ;;
 ;;                     ----==| T R A N S F O R M |==----                      ;;